Reorganise implementation into feature Modules
