+++ tmbr/Sources/App/Configurations/Configuration+Renderer.swift
@@ -1,0 +1,16 @@
+import Vapor
+import Core
+
+extension Configuration where Self == CoreConfiguration {
+    static var renderer: Self {
+        CoreConfiguration { app in
+            if app.environment == .production {
+                app.directory = DirectoryConfiguration(workingDirectory: "tmbr")
+            }
+            app.middleware.use(FileMiddleware(
+                publicDirectory: app.directory.publicDirectory
+            ))
+            app.views.use(.leaf)
+        }
+    }
+}
