+++ tmbr/Sources/App/Notifications/Web/Template+Notifications.swift
@@ -1,0 +1,5 @@
+import Core
+
+extension Template where Model == Never {
+    static let notifications = Template(name: "notifications")
+}
