+++ tmbr/Sources/App/Posts/Posts.swift
@@ -1,0 +1,21 @@
+import Fluent
+import Vapor
+import Core
+
+struct Posts: Module {
+    
+    func configure(_ app: Vapor.Application) throws {
+        app.migrations.add(CreatePost())
+    }
+    
+    func boot(_ app: Vapor.Application) throws {
+        try app.register(collection: PostsAPIController())
+        try app.register(collection: PostsWebController())
+    }
+}
+
+extension Module where Self == Posts {
+    static var posts: Self {
+        Posts()
+    }
+}
