--- tmbr/Sources/App/Routes/RSS/RSSViewModel.swift
@@ -1,20 +1,0 @@
-import Foundation
-
-struct RSSViewModel: Encodable {
-    
-    struct Post: Encodable {
-        let title: String
-        
-        let url: String
-        
-        let publishDate: String
-    }
-    
-    let title: String
-    
-    let url: String
-    
-    let description: String
-
-    let posts: [Post]
-}
