--- tmbr/Sources/App/Routes/Web/Auth/AuthController.swift
@@ -1,7 +1,0 @@
-import Vapor
-
-struct AuthController: RouteCollection {
-    func boot(routes: RoutesBuilder) throws {
-        routes.get("signin", page: .signin)
-    }
-}
