--- tmbr/Sources/App/Routes/Web/Posts/PostsController.swift
@@ -1,12 +1,0 @@
-import Vapor
-import Fluent
-
-struct PostsController: RouteCollection {
-    func boot(routes: RoutesBuilder) throws {
-        routes.get(page: .posts)
-        routes.get("post", ":postID", page: .post)
-
-        let protected = routes.grouped(User.redirectMiddleware(path: "/signin"))
-        protected.get("drafts", page: .drafts)
-    }
-}
