+++ tmbr/Sources/App/entrypoint.swift
@@ -2,6 +2,7 @@ import Vapor
 import Logging
 import NIOCore
 import NIOPosix
+import Core
 
 @main
 enum Entrypoint {
@@ -11,15 +12,24 @@ enum Entrypoint {
         
         let app = try await Application.make(env)
 
-        // This attempts to install NIO as the Swift Concurrency global executor.
-        // You can enable it if you'd like to reduce the amount of context switching between NIO and Swift Concurrency.
-        // Note: this has caused issues with some libraries that use `.wait()` and cleanly shutting down.
-        // If enabled, you should be careful about calling async functions before this point as it can cause assertion failures.
-        // let executorTakeoverSuccess = NIOSingletons.unsafeTryInstallSingletonPosixEventLoopGroupAsConcurrencyGlobalExecutor()
-        // app.logger.debug("Tried to install SwiftNIO's EventLoopGroup as Swift's global concurrency executor", metadata: ["success": .stringConvertible(executorTakeoverSuccess)])
+        let registry = ModuleRegistry(
+            configurations: [
+                .database,
+                .renderer,
+            ],
+            modules: [
+                .rss,
+                .manifest,
+                .authentication,
+                .notifications,
+                .posts,
+            ]
+        )
         
         do {
-            try await configure(app)
+            try registry.configure(app)
+            // try app.autoMigrate().wait()
+            try registry.boot(app)
         } catch {
             app.logger.report(error: error)
             try? await app.asyncShutdown()
