+++ tmbr/Sources/Core/Module/Module.swift
@@ -1,0 +1,5 @@
+import Vapor
+
+public protocol Module: Configuration {
+    func boot(_ app: Application) throws
+}
